video categories are working

+++ videos/views.py
@@ -1,11 +1,12 @@
-from django.shortcuts import render, redirect, get_object_or_404, HttpResponse
+from django.shortcuts import render, redirect, reverse, get_object_or_404, HttpResponse
 from django.contrib import messages
-from .models import Video
+from django.db.models import Q
+from .models import Video, VideoCategory
 
 
 def videos_home(request):
     videos = Video.objects.all()
-    my_video = get_object_or_404(Video, id=1)
+    my_video = get_object_or_404(Video, id=9)
     template = 'videos/videos_home.html'
     context = {
         'videos': videos,
@@ -15,6 +16,24 @@ def videos_home(request):
 
 def all_videos(request):
     videos = Video.objects.all()
+    if request.GET:
+        if 'category' in request.GET:
+            categories = request.GET['category'].split(',')
+            
+            videos = videos.filter(video_category__video_category__in=categories)         
+            categories = VideoCategory.objects.filter(video_category__in=categories)
+
+            if 'q' in request.GET:
+                query = request.GET['q']
+                if not query:
+                    messages.error(request, "You didn't enter any search criteria")
+                    return redirect(reverse('videos_home'))        
+
+                queries = Q(video_name__icontains=query) | Q(video_description__icontains=query)
+                products = products.filter(queries)
+                messages.success(request, f'{videos.count()} results found.')
+
+    
     context = {
         'videos': videos,
     }
@@ -27,7 +46,8 @@ def delete_video(request, video_id):
     if not request.user.is_superuser:
         messages.error(request, 'Sorry only store owners can do that')
         return redirect(reverse('RR_home'))
-    video = get_object_or_404(Video, pk=video_id)
-    video.delete()
+    if request.method == 'POST':
+        video = get_object_or_404(Video, pk=video_id)
+        video.delete()
     messages.success(request, 'Product deleted successfully!')
     return redirect('all_videos')
